+++ myproject/src/components/Dashboard.vue
@@ -2,12 +2,10 @@
   <div class="container">
     <img alt="Vue logo" src="../assets/logo.png">
     <br>
-    <h3><span>{{currentUser}}さんようこそ!</span>残高: {{wallet}}</h3>
+    <h3><span>{{ currentUser }}さんようこそ!</span>残高: {{ wallet }}</h3>
   </div>
 </template>
 <script>
-import 'firebase/auth';
-import 'firebase/firestore';
 export default {
   computed: {
     currentUser() {
