+++ myproject/src/components/Register.vue
@@ -18,9 +18,6 @@
 </div>
 </template>
 <script>
-import firebase from 'firebase/app';
-import "firebase/auth";
-import 'firebase/firestore';
 export default {
   data() {
     return {
@@ -31,26 +28,7 @@ export default {
   },
   methods: {
     register() {
-      firebase.auth().createUserWithEmailAndPassword(this.email, this.password) 
-      .then(() => {
-        firebase.auth().currentUser.updateProfile({
-          displayName: this.userName
-        }).then(() =>{
-          this.$router.push('/dashboard')
-        }).catch((error) => {
-          console.log(error);
-        })
-        const db = firebase.firestore()
-        this.$store.commit('setUserName', this.userName);
-        this.$store.commit('setWallet', 500)
-        db.collection('users').doc(this.userName).set(this.$store.getters);
-      })
-      .catch(error => {
-        console.log(error);
-        this.userName = '';
-        this.email = '';
-        this.password = '';
-      })
+      this.$store.dispatch('register', {userName: this.userName, email: this.email, password: this.password})
     }
   }  
 }
